@@ -5,21 +5,21 @@
 	<groupId>com.test</groupId>
 	<artifactId>zSpringBootTest</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<packaging>war</packaging>
+	<packaging>jar</packaging>
 	<name>zSpringBootTest</name>
 	<description>Spring Boot app for Testing purpose</description>
 
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.0</version>
+		<version>3.2.0</version>
 		<relativePath />
 	</parent>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 		<start-class>com.test.AppStarter</start-class>
 	</properties>
 	<dependencies>
@@ -56,12 +56,12 @@
 			<artifactId>json</artifactId>
 			<version>20180813</version>
 		</dependency>
-		<!--	To use XmlMapper	-->
-		<dependency>
-			<groupId>com.fasterxml.jackson.dataformat</groupId>
-			<artifactId>jackson-dataformat-xml</artifactId>
-<!--			<version>2.18.0</version>-->
-		</dependency>
+		<!--	To use XmlMapper : Convert reponse into XML	-->
+<!--		<dependency>-->
+<!--			<groupId>com.fasterxml.jackson.dataformat</groupId>-->
+<!--			<artifactId>jackson-dataformat-xml</artifactId>-->
+<!--&lt;!&ndash;			<version>2.18.0</version>&ndash;&gt;-->
+<!--		</dependency>-->
 
 	</dependencies>
 
