@@ -8,9 +8,8 @@ public class Test {
         System.out.println("------------Java-Program-Start-------------");
         int[] array = {4, 2, 3, 2, 1, 5};
         System.out.println("Original Array: ");
-        Arrays.stream(array).forEach(a -> System.out.print(a +" "));
-        System.out.println("\nSorted Array: ");
         Arrays.stream(array).sorted().forEach(a -> System.out.print(a +" "));
+
 
     }
 }
